[Rename] event : event 폴더 생성 후 위치 수정
