+++ coupon/src/main/java/com/haot/coupon/application/dto/request/events/EventModifyRequest.java
@@ -1,4 +1,4 @@
-package com.haot.coupon.application.dto.request;
+package com.haot.coupon.application.dto.request.events;
 
 import com.haot.coupon.domain.model.enums.EventStatus;
